requirements.txt

flask
gunicorn
python-dotenv
